proxyapi: Add Region field and WithRegion ctx
